Fix: rescoped wit command to be out of Execute() command handler method, moved tests to dedicated package.
